requirements.txt
streamlit
ultralytics
tensorflow
opencv-python
numpy
Pillow
scikit-learn